Converted DynamicArray<uint8> to a new data structure MemoryBuffer. This represents data that could be greater in size than uint32_max.
Created a stress test for my engine using various models. Current have less than 2000, no culling, no instancing, no batching, and it gets put down to 30 fps. This is due to my rendering implementation mostly. Update goes up a little, but mostly the rendering is the issue. All objects in the scene are moving in some way

+++ Source/EngineCore/Serialization/DeserializeBase.cpp
@@ -57,7 +57,7 @@ void Deserialize(DeserializeBase& ser, double& d)
 	ser.DeserializeData(&d, sizeof(double));
 }
 
-void Deserialize(DeserializeBase& ser, tchar*& c, uint32& size)
+void Deserialize(DeserializeBase& ser, tchar*& c, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
 	ser.DeserializeData(&size, sizeof(uint32));
@@ -65,82 +65,82 @@ void Deserialize(DeserializeBase& ser, tchar*& c, uint32& size)
 	ser.DeserializeData(c, size);
 }
 
-void Deserialize(DeserializeBase& ser, uint8*& u8, uint32& size)
+void Deserialize(DeserializeBase& ser, uint8*& u8, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	u8 = new uint8[size];
 	ser.DeserializeData(u8, size);
 }
 
-void Deserialize(DeserializeBase& ser, int8*& i8, uint32& size)
+void Deserialize(DeserializeBase& ser, int8*& i8, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	i8 = new int8[size];
 	ser.DeserializeData(i8, size);
 }
 
-void Deserialize(DeserializeBase& ser, uint16*& u16, uint32& size)
+void Deserialize(DeserializeBase& ser, uint16*& u16, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	u16 = new uint16[size];
 	ser.DeserializeData(u16, size * sizeof(uint16));
 }
 
-void Deserialize(DeserializeBase& ser, int16*& i16, uint32& size)
+void Deserialize(DeserializeBase& ser, int16*& i16, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	i16 = new int16[size];
 	ser.DeserializeData(i16, size * sizeof(int16));
 }
 
-void Deserialize(DeserializeBase& ser, uint32*& u32, uint32& size)
+void Deserialize(DeserializeBase& ser, uint32*& u32, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	u32 = new uint32[size];
 	ser.DeserializeData(u32, size * sizeof(uint32));
 }
 
-void Deserialize(DeserializeBase& ser, int32*& i32, uint32& size)
+void Deserialize(DeserializeBase& ser, int32*& i32, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	i32 = new int32[size];
 	ser.DeserializeData(i32, size * sizeof(int32));
 }
 
-void Deserialize(DeserializeBase& ser, uint64*& u64, uint32& size)
+void Deserialize(DeserializeBase& ser, uint64*& u64, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	u64 = new uint64[size];
 	ser.DeserializeData(u64, size * sizeof(uint64));
 }
 
-void Deserialize(DeserializeBase& ser, int64*& i64, uint32& size)
+void Deserialize(DeserializeBase& ser, int64*& i64, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	i64 = new int64[size];
 	ser.DeserializeData(i64, size * sizeof(int64));
 }
 
-void Deserialize(DeserializeBase& ser, float*& f, uint32& size)
+void Deserialize(DeserializeBase& ser, float*& f, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	f = new float[size];
 	ser.DeserializeData(f, size * sizeof(float));
 }
 
-void Deserialize(DeserializeBase& ser, double*& d, uint32& size)
+void Deserialize(DeserializeBase& ser, double*& d, size_t& size)
 {
 	// TODO - This allocates memory when the memory should be already allocated honestly...
-	ser.DeserializeData(&size, sizeof(uint32));
+	ser.DeserializeData(&size, sizeof(size_t));
 	d = new double[size];
 	ser.DeserializeData(d, size * sizeof(double));
 }
