+++ Source/EngineCore/Serialization/SerializeBase.hpp
@@ -10,7 +10,7 @@ public:
 	SerializeBase() = default;
 	virtual ~SerializeBase() = default;
 
-	virtual void SerializeData(const void* data, uint32 dataSize) = 0;
+	virtual void SerializeData(const void* data, size_t dataSize) = 0;
 
 	friend void Serialize(SerializeBase& ser, tchar c);
 	friend void Serialize(SerializeBase& ser, uint8 u16);
@@ -25,15 +25,15 @@ public:
 	friend void Serialize(SerializeBase& ser, double d);
 
 	// Raw pointer data serialization
-	friend void Serialize(SerializeBase& ser, const tchar* c, uint32 size);
-	friend void Serialize(SerializeBase& ser, const uint8* u16, uint32 size);
-	friend void Serialize(SerializeBase& ser, const int8* i16, uint32 size);
-	friend void Serialize(SerializeBase& ser, const uint16* u16, uint32 size);
-	friend void Serialize(SerializeBase& ser, const int16* i16, uint32 size);
-	friend void Serialize(SerializeBase& ser, const uint32* u32, uint32 size);
-	friend void Serialize(SerializeBase& ser, const int32* i32, uint32 size);
-	friend void Serialize(SerializeBase& ser, const uint64* u64, uint32 size);
-	friend void Serialize(SerializeBase& ser, const int64* i64, uint32 size);
-	friend void Serialize(SerializeBase& ser, const float* f, uint32 size);
-	friend void Serialize(SerializeBase& ser, const double* d, uint32 size);
+	friend void Serialize(SerializeBase& ser, const tchar* c, size_t size);
+	friend void Serialize(SerializeBase& ser, const uint8* u16, size_t size);
+	friend void Serialize(SerializeBase& ser, const int8* i16, size_t size);
+	friend void Serialize(SerializeBase& ser, const uint16* u16, size_t size);
+	friend void Serialize(SerializeBase& ser, const int16* i16, size_t size);
+	friend void Serialize(SerializeBase& ser, const uint32* u32, size_t size);
+	friend void Serialize(SerializeBase& ser, const int32* i32, size_t size);
+	friend void Serialize(SerializeBase& ser, const uint64* u64, size_t size);
+	friend void Serialize(SerializeBase& ser, const int64* i64, size_t size);
+	friend void Serialize(SerializeBase& ser, const float* f, size_t size);
+	friend void Serialize(SerializeBase& ser, const double* d, size_t size);
 };
